refactor : 테스트 코드 위치를 실제 소스코드 디렉토리 위치랑 일치하도록 변경
